More #70 OSX rake file changes
* they do work for targets mavericks and xsnow
* still not pretty in env.rb
* Commented out out, moved most OSX stuff from Rakefile to env.rb(s)

+++ make/darwin/tasks.rb
@@ -156,7 +156,7 @@ class MakeDarwin
       dylibs.each do |libn| 
         keyf = File.basename libn
         if !dups[keyf] 
-          cp "#{libn}", "#{TGT_DIR}/"
+          cp "#{libn}", "#{TGT_DIR}/" unless File.exists? "#{TGT_DIR}/#{keyf}"
           dups[keyf] = true
           chmod 0755, "#{TGT_DIR}/#{keyf}" unless File.writable? "#{TGT_DIR}/#{keyf}"
         end
